big_world_1939.zip
This is a compressed file folder containing all of the map files for a map I created, Big World 1939. This map is based on the Big World 1942 map that comes included with TripleA. This map is set in 1939, however, and has many more players: Germans, Russians, Slovaks, British, Romanians, French, Italians, Canadians, Japanese, Chinese, Finns, and Americans. Note that the Americans, though not officially in the war until 1942, still had the option to join the war. This is why I included them as a player.